fix: Timetable 분기 로직 hook 분리
