+++ src/pages/TimetablePage/index.tsx
@@ -7,22 +7,11 @@ import LectureList from 'components/TimetablePage/DefaultPage/LectureList';
 import MyLectureTimetable from 'components/TimetablePage/DefaultPage/MyLectureTimetable';
 import MyLectureList from 'components/TimetablePage/DefaultPage/MyLectureList';
 import Curriculum from 'components/TimetablePage/DefaultPage/Curriculum';
-import { useRecoilValue } from 'recoil';
-import { myLecturesAtom, selectedSemesterAtom } from 'utils/recoil/semester';
-import useTokenState from 'utils/hooks/useTokenState';
-import useTimetableInfoList from 'components/TimetablePage/hooks/useTimetableInfoList';
 import styles from './TimetablePage.module.scss';
 
 function TimetablePage() {
   const isMobile = useMediaQuery();
   useScrollToTop();
-
-  const token = useTokenState();
-  const selectedSemester = useRecoilValue(selectedSemesterAtom);
-  const myLecturesFromLocalStorageValue = useRecoilValue(myLecturesAtom);
-  const { data: myLecturesFromServer } = useTimetableInfoList(selectedSemester, token);
-
-  const myLectures = token ? (myLecturesFromServer ?? []) : (myLecturesFromLocalStorageValue ?? []);
 
   return (
     <div className={styles.page}>
@@ -31,17 +20,17 @@ function TimetablePage() {
           <h1 className={styles.page__title}>시간표</h1>
           <div className={styles.page__content}>
             {/* 강의 목록 */}
-            <LectureList myLectures={myLectures} />
+            <LectureList />
             {/* 나의 시간표 타임 테이블 */}
-            <MyLectureTimetable myLectures={myLectures} />
+            <MyLectureTimetable />
             {/* 나의 시간표 강의 목록 */}
-            <MyLectureList myLectures={myLectures} />
+            <MyLectureList />
             {/* 시간표 커리큘럼 */}
             <Curriculum />
           </div>
         </>
       ) : (
-        <MobilePage myLectures={myLectures} />
+        <MobilePage />
       ) }
     </div>
   );
